Updated Visual Studio project files.

+++ simple/util/condition_utils.cpp
@@ -78,7 +78,7 @@ class ConditionPrinter : public ConditionVisitor {
 
     void visit_pattern_condition(PatternCondition *condition) {
         _result = "(PatternCondition ";
-        _result += std::to_string((long) condition->get_expr_ast());
+        _result += std::to_string((long long) condition->get_expr_ast());
         _result += ")";
     }
